+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # JROFBridge CHANGELOG
 
+## 0.2.1
+
+* Added `-[JRProject remainingTasks]` - only returns those tasks which aren't complete.
+* Added `-[JRProject deferralType]` and `-[JRProject deferralLabel]` - determine whether the project or its first task (or neither) are deferred.
+
 ## 0.2.0
 
 Initial release *to public*. Exciting stability improvements, overall more flexible code, even closures in some places.